Django application/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.11.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
cryptography==39.0.0
defusedxml==0.7.1
diff-match-patch==20200713
Django==3.2.16
django-allauth==0.45.0
django-bootstrap==0.2.4
django-bootstrap4==22.3
django-bootstrap5==22.2
django-extensions==3.2.1
django-fontawesome-5==1.0.18
django-import-export==3.0.2
django-jasmine==0.5.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
importlib-metadata==2.1.3
joblib==1.2.0
MarkupPy==1.14
numpy==1.21.6
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.0.10
pandas==1.3.5
Pillow==9.3.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2022.7
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
scikit-learn==1.0.2
scipy==1.7.3
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tablib==3.3.0
threadpoolctl==3.1.0
typing-extensions==4.4.0
urllib3==1.26.14
xlrd==2.0.1
xlwt==1.3.0
zipp==3.11.0
